Lopading state and warning text while downlading
